backend/requirements.txt
# LLM
anthropic>=0.40.0
openai>=1.50.0

# ベクトルDB
qdrant-client>=1.12.0

# 非同期HTTP
httpx>=0.27.0

# キャッシュ
redis>=5.0.0

# 環境変数
python-dotenv>=1.0.0

# データ処理
numpy>=1.26.0
pandas>=2.2.0

# XML解析（EDINET有報）
lxml>=5.3.0

# 株価データ（バックテスト用）
yfinance>=0.2.40

# ユーティリティ
tenacity>=8.3.0      # リトライロジック
tqdm>=4.66.0         # プログレスバー
pydantic>=2.8.0      # バリデーション

# Web API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0

# テスト
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
